+ fix resubmit action + big switch

+++ src/main/scala/org/change/parser/p4/HeaderDeclaration.scala
@@ -6,10 +6,10 @@ package org.change.parser.p4
 case class HeaderDeclaration (
                                headerName: String,
                                // Map from header offset to header name and width
-                               fields: Map[Int, (String, Int)],
-                               length: Int
+                               fields: Map[Long, (String, Long)],
+                               length: Long
                              ) {
   // Shortcut to get width by name
-  lazy val indexOf: Map[String, Int] = fields.values.toMap
-  lazy val offsetOf: Map[String, Int] = fields.map(of => of._2._1 -> of._1)
+  lazy val indexOf: Map[String, Long] = fields.values.toMap
+  lazy val offsetOf: Map[String, Long] = fields.map(of => of._2._1 -> of._1)
 }
